Split DrrrClient into multiple files
- Split into property definitions, room interaction and site interaction.
240 Lines of code in one file already? I suggest this as a way to keep it a little more organised.
